Move spotify tests, fix google_ imports
